include/VBox/VBoxGuest.h: Solaris build fix for a narrowing conversion identified by gcc 12.1.0: VBGL_IOCTL_CODE_SIZE expands to _IOWRN which casts its result to an 'int' but is passed to routines expecting a 'uintptr_t'.  gcc correctly flags this via [-Wnarrowing] with an error message like: "narrowing conversion of ‘-1072146881’ from ‘int’ to ‘long unsigned int’".
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@96325 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/VBox/VBoxGuest.h
@@ -125,7 +125,7 @@
 #elif defined(RT_OS_SOLARIS)
   /* No automatic buffering, size limited to 255 bytes => use VBGLBIGREQ for everything. */
 # include <sys/ioccom.h>
-# define VBGL_IOCTL_CODE_SIZE(Function, Size)       _IOWRN('V', (Function), sizeof(VBGLREQHDR))
+# define VBGL_IOCTL_CODE_SIZE(Function, Size)       ((uintptr_t)(_IOWRN('V', (Function), sizeof(VBGLREQHDR))))
 # define VBGL_IOCTL_CODE_BIG(Function)              _IOWRN('V', (Function), sizeof(VBGLREQHDR))
 # define VBGL_IOCTL_CODE_FAST(Function)             _IO(   'F', (Function))
 # define VBGL_IOCTL_CODE_STRIPPED(a_uIOCtl)         (a_uIOCtl)
